Change calculation of target moves in distribution process.
Previously we created a distribution line for each target move but this meant
that we could put two lines with the same target production which is not
appropriate.

So we fix that situation by grouping quantities per product and production.

As we're using an SQL query to compute that, instead of searching for stock
moves for each purchase move this is also probably faster for large
distributions.